chore: Refactor deails  components to order
